app/rag/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
qdrant-client==1.8.0
sentence-transformers==2.5.1
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.27.0
sse-starlette==2.0.0
